refactor: move AvatarImage into SharedComponents
