+++ src/components/BioCardList/BioCardListItem/BioCardListItem.tsx
@@ -1,12 +1,12 @@
 import { FC, MutableRefObject } from "react";
 
 import useClickableCard from "../../../hooks/useClickableCard";
-import AvatarImage from "../../AvatarImage";
 import { BioData } from "../../BioModal/BioModal";
 import SocialButtons from "../../SocialButtons";
 import BoxBorders from "../../SpriteSheet/BrushSvgs/BoxBorders";
 import { Heading, P } from "../../Typography";
 
+import AvatarImage from "@/components/SharedComponents/AvatarImage";
 import Button from "@/components/SharedComponents/Button";
 import Flex from "@/components/SharedComponents/Flex";
