+++ src/components/SharedComponents/AvatarImage/AvatarImage.stories.tsx
