+++ src/components/SharedComponents/AvatarImage/index.ts
